Allow arbitrary objects as table row values
This allows both TypeScript `interfaces` without an index key and
objects with nested objects or arrays to be used in the
`input.table`/`select.table` IO methods.
This is useful particularly for `io.select.table` because it allows
receiving the full type of the input data for further operations.

Ideally, the user will use `columns` to serialize their data into
a flat table if their data contains nested objects, but if not we simply
send it through and display it as if it were serialized, which is
both useful if they don't particularly care about the output formatting,
and obviously signifies what's happening if they do and want to
customize it.

Closes #353