Add display.table partial data sending, refactor handle_state_change
In order to have the types work well without needing many more model
types for partials, we now pass the current state into the
`handle_state_change` callback and modify that before returning it.

This is less performant, we're basically converting to and from models
and dicts a couple times, but it may help us catch type errors;
correctness seems more important for now and we can optimize in the
future.