Cache pending page layouts, resend on new client connection
We have been treating pages as entirely stateless, which means we
basically recreate everything whenever a page is opened (which mimics
the web, mostly). This breaks down a bit when the same user connects to
the same page via multiple methods (fallback WebSocket and WebRTC
DataChannels).

We now perform caching of pending pages just like we
already do for transaction IO calls, and simply resend them when a new
client connects to an existing page session, or when an in-progress page
session reconnects. For python, where there are currently no peer
connections, we do the same but only resend on reconnection.

Fixes T-574