Fix loading clobbering pending IO calls
Because of `ctx.loading` call batching, it's not difficult to
accidentally have a loading state get sent immediately after an IO call,
which then renders the IO call invisible behind the new loading state.
To alleviate this, we now keep track of whether we have a pending IO
call in progress which is awaiting input, and prevent the loading state
from displaying if so.

cc https://github.com/interval/interval-node/issues/24