Add undefined support to python superjson, use for pending metadata
We now determine between values which are defined with the value of
`undefined`, and values that are not defined at all, and use that to
determine when to show a loading skeleton for the metadata values.
This is done using a special custom marker class Undefined in python
which is supported by our superjson implementation.